refactor(runtime): clean and better file architecture
