+++ packages/runtime/tests/disallow_codegen.rs
@@ -2,8 +2,8 @@ use std::sync::Once;
 
 use lagon_runtime::{
     http::{Request, RunResult},
-    isolate::{Isolate, IsolateOptions},
-    runtime::{Runtime, RuntimeOptions},
+    isolate::{options::IsolateOptions, Isolate},
+    runtime::{options::RuntimeOptions, Runtime},
 };
 
 fn setup() {
