+++ packages/runtime/tests/fetch.rs
@@ -3,8 +3,8 @@ use std::sync::Once;
 use httptest::{matchers::*, responders::*, Expectation, Server};
 use lagon_runtime::{
     http::{Request, Response, RunResult},
-    isolate::{Isolate, IsolateOptions},
-    runtime::{Runtime, RuntimeOptions},
+    isolate::{options::IsolateOptions, Isolate},
+    runtime::{options::RuntimeOptions, Runtime},
 };
 
 fn setup() {
