+++ packages/runtime/tests/streams.rs
@@ -4,8 +4,8 @@ use httptest::{matchers::*, responders::*, Expectation, Server};
 use hyper::body::Bytes;
 use lagon_runtime::{
     http::{Request, Response, RunResult, StreamResult},
-    isolate::{Isolate, IsolateOptions},
-    runtime::{Runtime, RuntimeOptions},
+    isolate::{options::IsolateOptions, Isolate},
+    runtime::{options::RuntimeOptions, Runtime},
 };
 
 fn setup() {
